requirements.txt
Flask==1.0.3
torchvision==0.2.1
numpy==1.16.4
Pillow==6.2.0
gunicorn==19.9.0
h5py==2.10.0
